+++ extensions/native/circuit/cuda/src/castf.cu
@@ -1,9 +1,8 @@
-#include "constants.h"
-#include "adapters/convert_adapter.cuh"
-#include "histogram.cuh"
 #include "launcher.cuh"
-#include "trace_access.h"
-#include "buffer_view.cuh"
+#include "native/adapters/convert_adapter.cuh"
+#include "primitives/buffer_view.cuh"
+#include "primitives/constants.h"
+#include "primitives/trace_access.h"
 
 using namespace riscv;
 using namespace program;
@@ -52,8 +51,8 @@ template <typename F> struct CastFRecord {
 
 __global__ void castf_tracegen(
     Fp *trace,
-    uint32_t height,
-    uint32_t width,
+    size_t height,
+    size_t width,
     DeviceBufferConstView<CastFRecord<Fp>> records,
     uint32_t *range_checker_ptr,
     uint32_t range_checker_num_bins,
@@ -62,11 +61,10 @@ __global__ void castf_tracegen(
     uint32_t idx = blockIdx.x * blockDim.x + threadIdx.x;
     RowSlice row(trace + idx, height);
     if (idx < records.len()) {
-        auto const& record = records[idx];
+        auto const &record = records[idx];
 
         auto adapter = ConvertAdapter<Fp, RV32_REGISTER_NUM_LIMBS>(
-            VariableRangeChecker(range_checker_ptr, range_checker_num_bins),
-            timestamp_max_bits
+            VariableRangeChecker(range_checker_ptr, range_checker_num_bins), timestamp_max_bits
         );
         adapter.fill_trace_row(row, record.adapter);
 
@@ -80,8 +78,8 @@ __global__ void castf_tracegen(
 
 extern "C" int _castf_tracegen(
     Fp *d_trace,
-    uint32_t height,
-    uint32_t width,
+    size_t height,
+    size_t width,
     DeviceBufferConstView<CastFRecord<Fp>> d_records,
     uint32_t *d_range_checker,
     uint32_t range_checker_num_bins,
@@ -91,8 +89,13 @@ extern "C" int _castf_tracegen(
     assert(width == sizeof(CastFCols<uint8_t>));
     auto [grid, block] = kernel_launch_params(height);
     castf_tracegen<<<grid, block>>>(
-        d_trace, height, width, d_records, d_range_checker, 
-        range_checker_num_bins, timestamp_max_bits
+        d_trace,
+        height,
+        width,
+        d_records,
+        d_range_checker,
+        range_checker_num_bins,
+        timestamp_max_bits
     );
     return cudaGetLastError();
 }
