+++ extensions/native/circuit/src/poseidon2/cuda.rs
@@ -2,7 +2,7 @@ use std::{borrow::Borrow, mem::size_of, slice::from_raw_parts, sync::Arc};
 
 use derive_new::new;
 use openvm_circuit::{arch::DenseRecordArena, utils::next_power_of_two_or_zero};
-use openvm_circuit_primitives::var_range::cuda::VariableRangeCheckerChipGPU;
+use openvm_circuit_primitives::var_range::VariableRangeCheckerChipGPU;
 use openvm_cuda_backend::{
     base::DeviceMatrix, chip::get_empty_air_proving_ctx, prover_backend::GpuBackend, types::F,
 };
@@ -72,10 +72,10 @@ impl<const SBOX_REGISTERS: usize> Chip<DenseRecordArena, GpuBackend>
         unsafe {
             poseidon2_cuda::tracegen(
                 trace.buffer(),
-                padded_height as u32,
-                width as u32,
+                padded_height,
+                width,
                 &d_records,
-                height as u32,
+                height,
                 &d_chunk_start,
                 d_chunk_start.len() as u32,
                 &self.range_checker.count,
